Restructure waste carriers pages
