+++ pages/submit-an-export/journey-of-waste/constants.ts
@@ -1,0 +1,15 @@
+export const links = {
+  wasteCarrier:
+    '/submit-an-export/journey-of-waste/waste-carriers/waste-carrier',
+  wasteCarrierTransport:
+    '/submit-an-export/journey-of-waste/waste-carriers/waste-carrier-transport',
+  wasteCarrierCheck:
+    '/submit-an-export/journey-of-waste/waste-carriers/waste-carrier-check',
+  wasteCarrierCollectContact:
+    '/submit-an-export/journey-of-waste/waste-collection-details/waste-carrier-collection-contact',
+  wasteCarrierCollectLocation:
+    '/submit-an-export/journey-of-waste/waste-collection-details/waste-carrier-collection-location',
+  wasteLeaveLocation: '/submit-an-export/journey-of-waste/waste-leave-location',
+  wasteTravelLocations:
+    '/submit-an-export/journey-of-waste/waste-travel-locations',
+};
