--- pages/submit-an-export/journey-of-waste/waste-carriers/constants.ts
@@ -1,16 +1,0 @@
-export const links = {
-  wasteCarrier:
-    '/submit-an-export/journey-of-waste/waste-carriers/waste-carrier',
-  wasteCarrierTransport:
-    '/submit-an-export/journey-of-waste/waste-carriers/waste-carrier-transport',
-  wasteCarrierCheck:
-    '/submit-an-export/journey-of-waste/waste-carriers/waste-carrier-check',
-  wasteCarrierCollect:
-    '/submit-an-export/journey-of-waste/waste-carriers/waste-carrier-collect',
-  wasteCarrierCollectDetails:
-    '/submit-an-export/journey-of-waste/waste-carriers/waste-carrier-collect-details',
-  wasteCarrierCollectLeave:
-    '/submit-an-export/journey-of-waste/waste-carriers/waste-carrier-collect-leave',
-  wasteCarrierCollectTravel:
-    '/submit-an-export/journey-of-waste/waste-carriers/waste-carrier-collect-travel',
-};
