+++ pages/submit-an-export/journey-of-waste/waste-collection-details/waste-carrier-collection-location/constants.ts
