+++ pages/submit-an-export/journey-of-waste/waste-leave-location/index.tsx
@@ -1,20 +1,20 @@
-import { Button, FormGroup, H2, HintText, Radio } from 'govuk-react';
+import { Button, FormGroup, H2, Radio } from 'govuk-react';
 import { useRouter } from 'next/router';
 import { Field, Form } from 'react-final-form';
-import { BaseTaskPage } from '../../../../../components/base-task-page';
-import { links } from '../../../../submit-export-task-list/constants';
+import { BaseTaskPage } from '../../../../components/base-task-page';
+import { links } from '../constants';
 import {
-  WASTE_CARRIER_COLLECT_TRAVEL_GROUP,
-  WASTE_CARRIER_COLLECT_TRAVEL_IDS,
-  WASTE_CARRIER_COLLECT_TRAVEL_LABELS,
+  WASTE_CARRIER_COLLECT_LEAVE_GROUP,
+  WASTE_CARRIER_COLLECT_LEAVE_IDS,
+  WASTE_CARRIER_COLLECT_LEAVE_LABELS,
 } from './constants';
 
-const WasteCarrierCollectTravel = () => {
+const WasteCarrierCollectLeave = () => {
   const router = useRouter();
 
   const handleSubmit = (data: any) => {
     console.log(data);
-    router.push(links.submitExportTaskList);
+    router.push(links.wasteTravelLocations);
   };
 
   return (
@@ -24,31 +24,28 @@ const WasteCarrierCollectTravel = () => {
           {({ handleSubmit }) => (
             <form onSubmit={handleSubmit}>
               <H2 size="LARGE">
-                Are there any other countries the waste will travel through?
+                Do you know the location at which the waste will leave the UK?
               </H2>
-              <HintText>
-                Do not include the exporting or importing country.
-              </HintText>
               <FormGroup>
                 <Field
-                  name={WASTE_CARRIER_COLLECT_TRAVEL_GROUP}
+                  name={WASTE_CARRIER_COLLECT_LEAVE_GROUP}
                   type="radio"
-                  value={WASTE_CARRIER_COLLECT_TRAVEL_IDS.YES}
+                  value={WASTE_CARRIER_COLLECT_LEAVE_IDS.YES}
                 >
                   {({ input }) => (
                     <Radio {...input}>
-                      {WASTE_CARRIER_COLLECT_TRAVEL_LABELS.YES}
+                      {WASTE_CARRIER_COLLECT_LEAVE_LABELS.YES}
                     </Radio>
                   )}
                 </Field>
                 <Field
-                  name={WASTE_CARRIER_COLLECT_TRAVEL_GROUP}
+                  name={WASTE_CARRIER_COLLECT_LEAVE_GROUP}
                   type="radio"
-                  value={WASTE_CARRIER_COLLECT_TRAVEL_IDS.NO}
+                  value={WASTE_CARRIER_COLLECT_LEAVE_IDS.NO}
                 >
                   {({ input }) => (
                     <Radio {...input}>
-                      {WASTE_CARRIER_COLLECT_TRAVEL_LABELS.NO}
+                      {WASTE_CARRIER_COLLECT_LEAVE_LABELS.NO}
                     </Radio>
                   )}
                 </Field>
@@ -64,4 +61,4 @@ const WasteCarrierCollectTravel = () => {
   );
 };
 
-export default WasteCarrierCollectTravel;
+export default WasteCarrierCollectLeave;
