+++ pages/submit-an-export/journey-of-waste/waste-travel-locations/index.tsx
@@ -1,20 +1,20 @@
-import { Button, FormGroup, H2, Radio } from 'govuk-react';
+import { Button, FormGroup, H2, HintText, Radio } from 'govuk-react';
 import { useRouter } from 'next/router';
 import { Field, Form } from 'react-final-form';
-import { BaseTaskPage } from '../../../../../components/base-task-page';
-import { links } from '../constants';
+import { BaseTaskPage } from '../../../../components/base-task-page';
+import { links } from '../../../submit-export-task-list/constants';
 import {
-  WASTE_CARRIER_COLLECT_LEAVE_GROUP,
-  WASTE_CARRIER_COLLECT_LEAVE_IDS,
-  WASTE_CARRIER_COLLECT_LEAVE_LABELS,
+  WASTE_CARRIER_COLLECT_TRAVEL_GROUP,
+  WASTE_CARRIER_COLLECT_TRAVEL_IDS,
+  WASTE_CARRIER_COLLECT_TRAVEL_LABELS,
 } from './constants';
 
-const WasteCarrierCollectLeave = () => {
+const WasteCarrierCollectTravel = () => {
   const router = useRouter();
 
   const handleSubmit = (data: any) => {
     console.log(data);
-    router.push(links.wasteCarrierCollectTravel);
+    router.push(links.submitExportTaskList);
   };
 
   return (
@@ -24,28 +24,31 @@ const WasteCarrierCollectLeave = () => {
           {({ handleSubmit }) => (
             <form onSubmit={handleSubmit}>
               <H2 size="LARGE">
-                Do you know the location at which the waste will leave the UK?
+                Are there any other countries the waste will travel through?
               </H2>
+              <HintText>
+                Do not include the exporting or importing country.
+              </HintText>
               <FormGroup>
                 <Field
-                  name={WASTE_CARRIER_COLLECT_LEAVE_GROUP}
+                  name={WASTE_CARRIER_COLLECT_TRAVEL_GROUP}
                   type="radio"
-                  value={WASTE_CARRIER_COLLECT_LEAVE_IDS.YES}
+                  value={WASTE_CARRIER_COLLECT_TRAVEL_IDS.YES}
                 >
                   {({ input }) => (
                     <Radio {...input}>
-                      {WASTE_CARRIER_COLLECT_LEAVE_LABELS.YES}
+                      {WASTE_CARRIER_COLLECT_TRAVEL_LABELS.YES}
                     </Radio>
                   )}
                 </Field>
                 <Field
-                  name={WASTE_CARRIER_COLLECT_LEAVE_GROUP}
+                  name={WASTE_CARRIER_COLLECT_TRAVEL_GROUP}
                   type="radio"
-                  value={WASTE_CARRIER_COLLECT_LEAVE_IDS.NO}
+                  value={WASTE_CARRIER_COLLECT_TRAVEL_IDS.NO}
                 >
                   {({ input }) => (
                     <Radio {...input}>
-                      {WASTE_CARRIER_COLLECT_LEAVE_LABELS.NO}
+                      {WASTE_CARRIER_COLLECT_TRAVEL_LABELS.NO}
                     </Radio>
                   )}
                 </Field>
@@ -61,4 +64,4 @@ const WasteCarrierCollectLeave = () => {
   );
 };
 
-export default WasteCarrierCollectLeave;
+export default WasteCarrierCollectTravel;
